Made get_term_meta return an array always
Looks like get_term_meta can return either a string or an array if the third argument is true. However, using false will force it to return an array even if it only finds 1 value.
This is what we need in the context of this code as the PassleTag constructor expects an optional array as a third argument and in no case, a string