Clean tag groups cache when option is deselected
When the user deselects the plugin option that creates custom taxonomies based on passle tag groups, we can clear the tag groups cache in the plugin.
This will ensure that the legacy value in this cache are not stored in the background and will make it easier for customers to just reset tag groups by toggling the relevant plugin option.