Changed OptionsService to make its get/set simpler
OptionsService get method was getting the default options and swapping values with the saved options. I changed it so it just returns the saved options. In the init() method, it will have set the saved options if there are no saved options already so we don't need to do this swapping all the time.
Also, the set method was updated to only do the setting of options if the second parameter of it is false.